Reposition Spatial Norm
